Crate cumulus_client_network

Parachain specific networking

Provides a custom block announcement implementation for parachains that use the relay chain provided consensus. See BlockAnnounceValidator and WaitToAnnounce for more information about this implementation.

Structs

BlockAnnounceData

The data that we attach to a block announcement.

BlockAnnounceValidator

Parachain specific block announce validator.

WaitToAnnounce

Wait before announcing a block that a candidate message has been received for this block, then add this message as justification for the block announcement.